+++ packages/staark-patch/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.0.1
+
+- Fix in children to node function.
+
 ## 1.0.0
 
 - Initial release.